requirements.txt
astronomer-cosmos>=1.0.2
astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy